requirements.txt
Flask==2.3.3
selenium==4.15.2
pytesseract==0.3.10
Pillow==10.0.1
fpdf2==2.7.6
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
webdriver-manager==4.0.1